Add files via upload
In our python boot camp, we created a bulk SMS sender as a project by using the services of twilio REST API's trial version. we used the learned concept of classes in our project, while learning the concept of abstraction while making the project. we abstracted the complex implementation details from the user as he can send the message only by inputting message text and the number to which sms is to be sent.